pyproject.toml
[tool.poetry]
name = "Rag_Application"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [
    { include = "app" },
]

[tool.poetry.dependencies]
python = "^3.11"
uvicorn = "^0.23.2"
langserve = {extras = ["server"], version = ">=0.0.30"}
pydantic = "^2.12.5"
requests = "^2.32.5"
tqdm = "^4.67.1"
unstructured = {extras = ["all-docs"], version = "^0.18.21"}
python-dotenv = "^1.2.1"
tiktoken = "^0.12.0"
pgvector = "^0.4.1"
langchain-core = "^1.1.0"
langchain-community = "^0.4.1"
langchain-experimental = "^0.4.0"
langchain-google-genai = "^3.2.0"
psycopg = {extras = ["binary"], version = "^3.2.13"}


[tool.poetry.group.dev.dependencies]
langchain-cli = ">=0.0.15"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
